Final solution
All functions implemented and tested. Debug task issues not resolved
(Invalid command is commented out, and backspace is not forwarded yet to
application).  "Dark mode" cursor blink not implemented.  Exit from Dark
mode is not smooth since LEDs may be in middle of system time for fade
or on/off, but this recovers fully on the next cycle.