Search timeout, OWER fixes, 32 messages
Added setting the ANT searching timeout so it starts at infinity for slaves.
Fixed the OWER configuration setting that caused ANT functionality to stop working.
Increased the debug message buffer from 16 to 32 to avoid filling the buffer.  There still should be a warning message if a message is added without space.