Fix AntExpectResponse
AntExpectResponse did not return success for requested messages.  It only returned success for successful event messages.  This became apparent when a debug message was printed after the version request message was sent and  AntExpectResponse is called.  Since this is the only instance now where AntExpectResponse is used, almost deleted the function.  However, it might be useful in the future, so fixed the behviour by checking first for an event, and if not an event, checking directly for the expected message in the response.